Consolidate client modules into single core
